x86_64 port: ALMOST working kernel. still fails and such

+++ kernel/gdt/gdt.h
@@ -18,14 +18,14 @@ extern gdt_entry_t gdt[3];
 
 struct gdt_ptr_struct {
     uint16_t limit;
-    uint32_t base;
+    uint64_t base;
 } __attribute__((packed));
 
 typedef struct gdt_ptr_struct gdt_ptr_t;
 
 extern gdt_ptr_t gdt_p; 
 
-void gdt_set_gate(int32_t num, uint32_t base, uint32_t limit, uint8_t access, uint8_t gran);
+void gdt_set_gate(int32_t num, uint64_t base, uint32_t limit, uint8_t access, uint8_t gran);
 
 void init_gdt();
